requirements.txt
absl-py==2.1.0
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
Farama-Notifications==0.0.4
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.3.0
grpcio==1.71.0
gymnasium==1.0.0
Jinja2==3.1.6
kiwisolver==1.4.8
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==6.30.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
setuptools==76.0.0
six==1.17.0
stable_baselines3==2.5.0
standard-imghdr==3.13.0
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
torch==2.6.0
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
websocket-client==1.8.0
Werkzeug==3.1.3
